+++ skyfield/api.py
@@ -10,19 +10,19 @@ from datetime import datetime
 from math import pi
 from .constants import tau
 from .errors import DeprecationError
-from .iokit import Loader
+from .iokit import Loader, load_file
 from .starlib import Star
 from .sgp4lib import EarthSatellite
 from .timelib import T0, Time, Timescale, utc
 from .toposlib import Topos
-from .units import Angle
+from .units import Angle, Distance
 from .named_stars import NamedStar
 
 load = Loader('.')
 
-__all__ = ['Angle', 'Loader', 'NamedStar', 'Star', 'EarthSatellite',
-           'T0', 'Time', 'Timescale', 'Topos', 'datetime',
-           'load', 'utc', 'pi', 'tau']
+__all__ = ['Angle', 'Distance', 'EarthSatellite', 'Loader', 'NamedStar',
+           'Star', 'T0', 'Time', 'Timescale', 'Topos', 'datetime',
+           'load', 'load_file', 'utc', 'pi', 'tau']
 
 # An attempt at friendliest-possible deprecations: