+++ skyfield/units.py
@@ -7,11 +7,14 @@ import numpy as np
 import sys
 from numpy import abs, array, copysign, isnan
 from .constants import AU_KM, AU_M, DAY_S, tau
+from .descriptorlib import reify
 
 def _to_array(value):
     """As a convenience, turn Python lists and tuples into NumPy arrays."""
     if isinstance(value, (tuple, list)):
         return array(value)
+    elif isinstance(value, (float, int)):
+        return np.float64(value)
     else:
         return value
 
@@ -39,21 +42,24 @@ class Distance(object):
         else:
             raise ValueError('to construct a Distance provide au, km, or m')
 
-    def __getattr__(self, name):
-        if name == 'km':
-            self.km = km = self.au * AU_KM
-            return km
-        if name == 'm':
-            self.m = m = self.au * AU_M
-            return m
-        if name == 'AU':
-            if not Distance._warned:
-                print('WARNING: the IAU has renamed the astronomical unit to'
-                      ' lowercase "au" so Skyfield will soon remove uppercase'
-                      ' "AU" from Distance objects', file=sys.stdout)
-                Distance._warned = True
-            return self.au
-        raise AttributeError('no attribute named %r' % (name,))
+    @reify
+    def km(self):
+        self.km = km = self.au * AU_KM
+        return km
+
+    @reify
+    def m(self):
+        self.m = m = self.au * AU_M
+        return m
+
+    @reify
+    def AU(self):
+        if not Distance._warned:
+            print('WARNING: the IAU has renamed the astronomical unit to'
+                  ' lowercase "au" so Skyfield will soon remove uppercase'
+                  ' "AU" from Distance objects', file=sys.stdout)
+            Distance._warned = True
+        return self.au
 
     def __str__(self):
         n = self.au
@@ -81,22 +87,25 @@ class Velocity(object):
     """
     _warned = False
 
-    def __init__(self, au_per_d):
+    def __init__(self, au_per_d=None, km_per_s=None):
+        if km_per_s is not None:
+            au_per_d = km_per_s * DAY_S / AU_KM
         self.au_per_d = _to_array(au_per_d)
 
-    def __getattr__(self, name):
-        if name == 'km_per_s':
-            self.km_per_s = self.au_per_d * AU_KM / DAY_S
-            return self.km_per_s
-        if name == 'AU_per_d':
-            if not Velocity._warned:
-                print('WARNING: the IAU has renamed the astronomical unit to'
-                      ' lowercase "au" so Skyfield will soon remove'
-                      ' "AU_per_day" in favor of "au_per_day"',
-                      file=sys.stdout)
-                Velocity._warned = True
+    @reify
+    def km_per_s(self):
+        self.km_per_s = self.au_per_d * AU_KM / DAY_S
+        return self.km_per_s
+
+    @reify
+    def AU_per_d(self):
+        if not Velocity._warned:
+            print('WARNING: the IAU has renamed the astronomical unit to'
+                  ' lowercase "au" so Skyfield will soon remove'
+                  ' "AU_per_day" in favor of "au_per_day"',
+                  file=sys.stdout)
+            Velocity._warned = True
             return self.au_per_d
-        raise AttributeError('no attribute named %r' % (name,))
 
     def __str__(self):
         n = self.au_per_d
@@ -142,7 +151,8 @@ Angle(radians=value)
 Angle(degrees=value)
 Angle(hours=value)
 
-where `value` can be either a Python float or a NumPy array of floats"""
+where `value` can be either a Python float, a list of Python floats, 
+or a NumPy array of floats"""
 
 class Angle(object):
 
@@ -154,12 +164,12 @@ class Angle(object):
                 raise ValueError(_instantiation_instructions)
             self.radians = angle.radians
         elif radians is not None:
-            self.radians = radians
+            self.radians = _to_array(radians)
         elif degrees is not None:
-            self._degrees = degrees = _unsexagesimalize(degrees)
+            self._degrees = degrees = _to_array(_unsexagesimalize(degrees))
             self.radians = degrees * _from_degrees
         elif hours is not None:
-            self._hours = hours = _unsexagesimalize(hours)
+            self._hours = hours = _to_array(_unsexagesimalize(hours))
             self.radians = hours * _from_hours
 
         self.preference = (preference if preference is not None
@@ -187,10 +197,15 @@ class Angle(object):
         raise AttributeError('no attribute named %r' % (name,))
 
     def __str__(self):
+        if self.radians.size == 0:
+            return 'Angle []'
         return self.dstr() if self.preference == 'degrees' else self.hstr()
 
     def __repr__(self):
-        return '<{0} {1}>'.format(type(self).__name__, self)
+        if self.radians.size == 0:
+            return '<{0} []>'.format(type(self).__name__, self)
+        else:
+            return '<{0} {1}>'.format(type(self).__name__, self)
 
     def hms(self, warn=True):
         """Convert to a tuple (hours, minutes, seconds).
@@ -218,8 +233,11 @@ class Angle(object):
         """Convert to a string like ``12h 07m 30.00s``."""
         if warn and self.preference != 'hours':
             raise WrongUnitError('hstr')
+        if self.radians.size == 0:
+            return '<Angle []>'
         hours = self._hours
-        if getattr(hours, 'shape', ()):
+        shape = getattr(hours, 'shape', ())
+        if shape and shape != (1,):
             return "{0} values from {1} to {2}".format(
                 len(hours),
                 _hstr(min(hours), places),
@@ -253,9 +271,12 @@ class Angle(object):
         """Convert to a string like ``181deg 52\' 30.0"``."""
         if warn and self.preference != 'degrees':
             raise WrongUnitError('dstr')
+        if self.radians.size == 0:
+            return '<Angle []>'
         degrees = self._degrees
         signed = self.signed
-        if getattr(degrees, 'shape', ()):
+        shape = getattr(degrees, 'shape', ())
+        if shape and shape != (1,):
             return "{0} values from {1} to {2}".format(
                 len(degrees),
                 _dstr(min(degrees), places, signed),